requirements.txt
tabulate
requests
pytest
python-dotenv
